Vision processing
added code to go through the various particles to test for size, test
the ratio of width and height and determine whether the particle is a
vertical or horizontal rectangle and determine whether the goal we are
viewing is hot or not. We also started to add parts to the
autosequencer.